Move the contributor docs to docs/ and the dev tooling to tools/
CONTRIBUTING, SECURITY, DESIGN and THIRD_PARTY_NOTICES now sit under docs/,
which GitHub still recognises for the first two. The eslint, playwright and
tailwind configs and the two check scripts move to tools/, with the paths
inside them made relative to the repository root rather than the config's
own directory. A one-line root eslint.config.mjs re-exports the real one so
a bare `npx eslint .` and editor integrations keep working.
