+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -6,5 +6,5 @@
 
 - [ ] `npm run check && npm run lint && npm test` pass locally
 - [ ] Tested in at least one desktop browser and one mobile browser (or explained why not)
-- [ ] No third-party code added without a compatible license noted in `THIRD_PARTY_NOTICES.md`
+- [ ] No third-party code added without a compatible license noted in `docs/THIRD_PARTY_NOTICES.md`
 - [ ] `CHANGELOG.md` updated if user-facing
