+++ changelog.html
@@ -67,7 +67,7 @@
 
   <footer class="cl-footer">
     <div class="cl-footer__inner">
-      <p>Zaya is released under the MIT licence. Third-party notices are listed in <code class="cl-code">THIRD_PARTY_NOTICES.md</code>.</p>
+      <p>Zaya is released under the MIT licence. Third-party notices are listed in <code class="cl-code">docs/THIRD_PARTY_NOTICES.md</code>.</p>
       <p><a class="cl-link" href="index.html">Open the reader</a></p>
     </div>
   </footer>
