+++ index.html
@@ -18,7 +18,7 @@
   <link rel="preload" href="lib/js/libs/pdf.min.js?v=6.2.0" as="script">
   <link rel="preload" href="lib/fonts/webfonts/fa-solid-900.woff2" as="font" type="font/woff2" crossorigin>
 
-  <!-- All third-party assets are vendored (see THIRD_PARTY_NOTICES.md); no runtime CDN dependency -->
+  <!-- All third-party assets are vendored (see docs/THIRD_PARTY_NOTICES.md); no runtime CDN dependency -->
   <link rel="stylesheet" href="lib/css/vendor/tailwind.css?v=6.2.0">
   <link rel="stylesheet" href="lib/css/vendor/fontawesome.min.css?v=6.2.0">
   <link rel="stylesheet" href="lib/css/vendor/toastify.min.css?v=6.2.0">
